requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
datasets>=2.14.0

# RAG and Vector Database
qdrant-client>=1.7.0
sentence-transformers>=2.2.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# API and utilities
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional but recommended
scikit-learn>=1.3.0
